Slim down the dummy test payment gateway
Follow-up to the removal of the deprecated Dummy plugins, treating what is
left purely as a test helper:

- Collapse the two gateways into one. The credit-card variant only added
  token-driven capture failures, so DummyGatewayPlugin absorbs them:
  capture() checks validate_token() first and falls back to dummy_success().
  Only two tests relied on that behaviour. The payment_dummy_credit_card
  fixture had become a byte-for-byte duplicate of payment_dummy and is gone.
- Drop the gateways' own test files. Every PaymentError they asserted is
  already covered against saleor/payment/utils.py by
  saleor/payment/tests/test_payment.py, test_dummy.py was a strict subset of
  test_dummy_credit_card.py, and the gateway orchestration is covered by
  test_gateway.py and the payment mutation tests.
- Drop DEFAULT_CONFIGURATION/CONFIG_STRUCTURE. No test configures this plugin
  through pluginUpdate, so GatewayConfig is now built inline from the values
  the tests actually depend on.
- Drop the unused pre-authorization tokens. "1112" was referenced nowhere and
  "1111" yields an identical result without the branch. The ChargeStatus
  token branch stays: the e2e order-void tests drive it with "not-charged".
