+++ saleor/graphql/checkout/tests/mutations/test_checkout_complete_with_payment.py
@@ -10,7 +10,7 @@ from django.db.models.aggregates import Sum
 from django.test import override_settings
 from django.utils import timezone
 
-from saleor.plugins.tests.gateways.dummy_credit_card import TOKEN_VALIDATION_MAPPING
+from saleor.plugins.tests.gateways.dummy import TOKEN_VALIDATION_MAPPING
 
 from .....account.models import Address
 from .....checkout import calculations
@@ -2981,18 +2981,13 @@ def test_checkout_complete_checkout_without_lines(
 
 @pytest.mark.integration
 @pytest.mark.parametrize(("token", "error"), list(TOKEN_VALIDATION_MAPPING.items()))
-@patch(
-    "saleor.plugins.tests.gateways.dummy_credit_card."
-    "DummyCreditCardGatewayPlugin.DEFAULT_ACTIVE",
-    True,
-)
-def test_checkout_complete_error_in_gateway_response_for_dummy_credit_card(
+def test_checkout_complete_error_in_gateway_response_for_declined_card(
     token,
     error,
     user_api_client,
     checkout_with_gift_card,
     gift_card,
-    payment_dummy_credit_card,
+    payment_dummy,
     address,
     checkout_delivery,
 ):
@@ -3015,7 +3010,7 @@ def test_checkout_complete_error_in_gateway_response_for_dummy_credit_card(
     total = calculations.calculate_checkout_total_with_gift_cards(
         manager, checkout_info, lines
     )
-    payment = payment_dummy_credit_card
+    payment = payment_dummy
     payment.is_active = True
     payment.order = None
     payment.total = total.gross.amount
@@ -4645,7 +4640,7 @@ def test_checkout_complete_error_when_shipping_address_doesnt_have_all_required_
     user_api_client,
     checkout_with_item,
     gift_card,
-    payment_dummy_credit_card,
+    payment_dummy,
     address,
     checkout_delivery,
 ):
@@ -4672,7 +4667,7 @@ def test_checkout_complete_error_when_shipping_address_doesnt_have_all_required_
     total = calculations.calculate_checkout_total_with_gift_cards(
         manager, checkout_info, lines
     )
-    payment = payment_dummy_credit_card
+    payment = payment_dummy
     payment.is_active = True
     payment.order = None
     payment.total = total.gross.amount
@@ -4707,7 +4702,7 @@ def test_checkout_complete_error_when_shipping_address_doesnt_have_all_valid_fie
     user_api_client,
     checkout_with_item,
     gift_card,
-    payment_dummy_credit_card,
+    payment_dummy,
     address,
     checkout_delivery,
 ):
@@ -4735,7 +4730,7 @@ def test_checkout_complete_error_when_shipping_address_doesnt_have_all_valid_fie
     total = calculations.calculate_checkout_total_with_gift_cards(
         manager, checkout_info, lines
     )
-    payment = payment_dummy_credit_card
+    payment = payment_dummy
     payment.is_active = True
     payment.order = None
     payment.total = total.gross.amount
@@ -4770,7 +4765,7 @@ def test_checkout_complete_error_when_billing_address_doesnt_have_all_required_f
     user_api_client,
     checkout_with_item,
     gift_card,
-    payment_dummy_credit_card,
+    payment_dummy,
     address,
     checkout_delivery,
 ):
@@ -4797,7 +4792,7 @@ def test_checkout_complete_error_when_billing_address_doesnt_have_all_required_f
     total = calculations.calculate_checkout_total_with_gift_cards(
         manager, checkout_info, lines
     )
-    payment = payment_dummy_credit_card
+    payment = payment_dummy
     payment.is_active = True
     payment.order = None
     payment.total = total.gross.amount
@@ -4831,7 +4826,7 @@ def test_checkout_complete_error_when_billing_address_doesnt_have_all_valid_fiel
     user_api_client,
     checkout_with_item,
     gift_card,
-    payment_dummy_credit_card,
+    payment_dummy,
     address,
     checkout_delivery,
 ):
@@ -4859,7 +4854,7 @@ def test_checkout_complete_error_when_billing_address_doesnt_have_all_valid_fiel
     total = calculations.calculate_checkout_total_with_gift_cards(
         manager, checkout_info, lines
     )
-    payment = payment_dummy_credit_card
+    payment = payment_dummy
     payment.is_active = True
     payment.order = None
     payment.total = total.gross.amount
