+++ saleor/plugins/tests/fixtures/gateway.py
@@ -2,11 +2,8 @@ import pytest
 
 
 @pytest.fixture(autouse=True)
-def setup_dummy_gateways(settings):
-    settings.PLUGINS = [
-        "saleor.plugins.tests.gateways.dummy.DummyGatewayPlugin",
-        "saleor.plugins.tests.gateways.dummy_credit_card.DummyCreditCardGatewayPlugin",
-    ]
+def setup_dummy_gateway(settings):
+    settings.PLUGINS = ["saleor.plugins.tests.gateways.dummy.DummyGatewayPlugin"]
     return settings
